changement page module categorie ok
